Add files via upload
This is a tasklist application that works on the command line and does not require any other external libraries. It is just a small holiday project which I completed in a day and just wanted to share for beginners to learn and also people who are interested can use this!